Added logic for generating Fibonacci sequence
First, the program prompts for the first 2 element and the number of elements to be generated.

Then, the program display the first two elements before proceeding to generate the next few elements by passing the element that comes before the next element and the limit to control the number of generated elements.

In the Fibonacci function, the recursion stops when the limit will be 0 before return the last element.